Rewrite message to struct conversions, add ReceiveMessageTransformer, for parsing long and forward messages in the future
