+++ mirai-core/src/commonMain/kotlin/message/incomingSourceImpl.kt
@@ -17,7 +17,7 @@ import net.mamoe.mirai.Bot
 import net.mamoe.mirai.contact.Friend
 import net.mamoe.mirai.contact.Member
 import net.mamoe.mirai.contact.Stranger
-import net.mamoe.mirai.internal.contact.GroupImpl
+import net.mamoe.mirai.internal.contact.checkIsGroupImpl
 import net.mamoe.mirai.internal.contact.newAnonymous
 import net.mamoe.mirai.internal.network.protocol.data.proto.ImMsgBody
 import net.mamoe.mirai.internal.network.protocol.data.proto.MsgComm
@@ -25,66 +25,32 @@ import net.mamoe.mirai.internal.network.protocol.data.proto.SourceMsg
 import net.mamoe.mirai.internal.network.protocol.packet.EMPTY_BYTE_ARRAY
 import net.mamoe.mirai.internal.utils._miraiContentToString
 import net.mamoe.mirai.internal.utils.io.serialization.toByteArray
-import net.mamoe.mirai.message.data.*
+import net.mamoe.mirai.message.data.MessageChain
+import net.mamoe.mirai.message.data.MessageSourceKind
+import net.mamoe.mirai.message.data.OnlineMessageSource
 import net.mamoe.mirai.utils.encodeToBase64
 import net.mamoe.mirai.utils.encodeToString
 import net.mamoe.mirai.utils.mapToIntArray
 import java.util.concurrent.atomic.AtomicBoolean
 
-internal interface MessageSourceInternal {
-    @Transient
-    val sequenceIds: IntArray // ids
-
-    @Transient
-    val internalIds: IntArray // randomId
-
-    @Deprecated("don't use this internally. Use sequenceId or random instead.", level = DeprecationLevel.ERROR)
-    @Transient
-    val ids: IntArray
-
-    @Transient
-    val isRecalledOrPlanned: AtomicBoolean
-
-    fun toJceData(): ImMsgBody.SourceMsg
-}
-
-@Suppress("RedundantSuspendModifier", "unused")
-internal suspend fun MessageSource.ensureSequenceIdAvailable() {
-    if (this is OnlineMessageSourceToGroupImpl) {
-        ensureSequenceIdAvailable()
-    }
-}
-
-@Suppress("RedundantSuspendModifier", "unused")
-internal suspend inline fun Message.ensureSequenceIdAvailable() {
-    (this as? MessageChain)?.sourceOrNull?.ensureSequenceIdAvailable()
-}
-
 @Serializable(OnlineMessageSourceFromFriendImpl.Serializer::class)
 internal class OnlineMessageSourceFromFriendImpl(
     override val bot: Bot,
-    val msg: List<MsgComm.Msg>
+    msg: List<MsgComm.Msg>
 ) : OnlineMessageSource.Incoming.FromFriend(), MessageSourceInternal {
     object Serializer : MessageSourceSerializerImpl("OnlineMessageSourceFromFriend")
 
-    override val sequenceIds: IntArray get() = msg.mapToIntArray { it.msgHead.msgSeq }
+    override val sequenceIds: IntArray = msg.mapToIntArray { it.msgHead.msgSeq }
     override var isRecalledOrPlanned: AtomicBoolean = AtomicBoolean(false)
     override val ids: IntArray get() = sequenceIds// msg.msgBody.richText.attr!!.random
-    override val internalIds: IntArray
-        get() = msg.mapToIntArray {
-            it.msgBody.richText.attr?.random ?: 0
-        } // other client 消息的这个是0
-    override val time: Int get() = msg.first().msgHead.msgTime
+    override val internalIds: IntArray = msg.mapToIntArray {
+        it.msgBody.richText.attr?.random ?: 0
+    } // other client 消息的这个是0
+    override val time: Int = msg.first().msgHead.msgTime
     override val originalMessage: MessageChain by lazy {
-        msg.toMessageChain(
-            bot,
-            bot.id,
-            0,
-            null,
-            MessageSourceKind.FRIEND
-        )
+        msg.toMessageChainNoSource(bot.id, 0, MessageSourceKind.FRIEND)
     }
-    override val sender: Friend get() = bot.getFriendOrFail(msg.first().msgHead.fromUin)
+    override val sender: Friend = bot.getFriendOrFail(msg.first().msgHead.fromUin)
 
     private val jceData by lazy { msg.toJceDataPrivate(internalIds) }
 
@@ -94,28 +60,21 @@ internal class OnlineMessageSourceFromFriendImpl(
 @Serializable(OnlineMessageSourceFromStrangerImpl.Serializer::class)
 internal class OnlineMessageSourceFromStrangerImpl(
     override val bot: Bot,
-    val msg: List<MsgComm.Msg>
+    msg: List<MsgComm.Msg>
 ) : OnlineMessageSource.Incoming.FromStranger(), MessageSourceInternal {
     object Serializer : MessageSourceSerializerImpl("OnlineMessageSourceFromStranger")
 
-    override val sequenceIds: IntArray get() = msg.mapToIntArray { it.msgHead.msgSeq }
+    override val sequenceIds: IntArray = msg.mapToIntArray { it.msgHead.msgSeq }
     override var isRecalledOrPlanned: AtomicBoolean = AtomicBoolean(false)
     override val ids: IntArray get() = sequenceIds// msg.msgBody.richText.attr!!.random
-    override val internalIds: IntArray
-        get() = msg.mapToIntArray {
-            it.msgBody.richText.attr?.random ?: 0
-        } // other client 消息的这个是0
-    override val time: Int get() = msg.first().msgHead.msgTime
+    override val internalIds: IntArray = msg.mapToIntArray {
+        it.msgBody.richText.attr?.random ?: 0
+    } // other client 消息的这个是0
+    override val time: Int = msg.first().msgHead.msgTime
     override val originalMessage: MessageChain by lazy {
-        msg.toMessageChain(
-            bot,
-            bot.id,
-            0,
-            null,
-            MessageSourceKind.STRANGER
-        )
+        msg.toMessageChainNoSource(bot.id, 0, MessageSourceKind.STRANGER)
     }
-    override val sender: Stranger get() = bot.getStrangerOrFail(msg.first().msgHead.fromUin)
+    override val sender: Stranger = bot.getStrangerOrFail(msg.first().msgHead.fromUin)
 
     private val jceData by lazy { msg.toJceDataPrivate(internalIds) }
 
@@ -126,100 +85,101 @@ private fun List<MsgComm.Msg>.toJceDataPrivate(ids: IntArray): ImMsgBody.SourceM
     val elements = flatMap { it.msgBody.richText.elems }.toMutableList().also {
         if (it.last().elemFlags2 == null) it.add(ImMsgBody.Elem(elemFlags2 = ImMsgBody.ElemFlags2()))
     }
-    return ImMsgBody.SourceMsg(
-        origSeqs = mapToIntArray { it.msgHead.msgSeq },
-        senderUin = first().msgHead.fromUin,
-        toUin = first().msgHead.toUin,
-        flag = 1,
-        elems = flatMap { it.msgBody.richText.elems },
-        type = 0,
-        time = this.first().msgHead.msgTime,
-        pbReserve = SourceMsg.ResvAttr(
-            origUids = ids.map { it.toLong() and 0xFFFF_FFFF }
-        ).toByteArray(SourceMsg.ResvAttr.serializer()),
-        srcMsg = MsgComm.Msg(
-            msgHead = MsgComm.MsgHead(
-                fromUin = this.first().msgHead.fromUin, // qq
-                toUin = this.first().msgHead.toUin, // group
-                msgType = this.first().msgHead.msgType, // 82?
-                c2cCmd = this.first().msgHead.c2cCmd,
-                msgSeq = this.first().msgHead.msgSeq,
-                msgTime = this.first().msgHead.msgTime,
-                msgUid = ids.single().toLong() and 0xFFFF_FFFF, // ok
-                // groupInfo = MsgComm.GroupInfo(groupCode = this.msgHead.groupInfo.groupCode),
-                isSrcMsg = true
-            ),
-            msgBody = ImMsgBody.MsgBody(
-                richText = ImMsgBody.RichText(
-                    elems = elements
+
+    first().msgHead.run {
+        return ImMsgBody.SourceMsg(
+            origSeqs = mapToIntArray { it.msgHead.msgSeq },
+            senderUin = fromUin,
+            toUin = toUin,
+            flag = 1,
+            elems = flatMap { it.msgBody.richText.elems },
+            type = 0,
+            time = msgTime,
+            pbReserve = SourceMsg.ResvAttr(
+                origUids = ids.map { it.toLong() and 0xFFFF_FFFF }
+            ).toByteArray(SourceMsg.ResvAttr.serializer()),
+            srcMsg = MsgComm.Msg(
+                msgHead = MsgComm.MsgHead(
+                    fromUin = fromUin, // qq
+                    toUin = toUin, // group
+                    msgType = msgType, // 82?
+                    c2cCmd = c2cCmd,
+                    msgSeq = msgSeq,
+                    msgTime = msgTime,
+                    msgUid = ids.single().toLong() and 0xFFFF_FFFF, // ok
+                    // groupInfo = MsgComm.GroupInfo(groupCode = msgHead.groupInfo.groupCode),
+                    isSrcMsg = true
+                ),
+                msgBody = ImMsgBody.MsgBody(
+                    richText = ImMsgBody.RichText(
+                        elems = elements
+                    )
                 )
-            )
-        ).toByteArray(MsgComm.Msg.serializer())
-    )
+            ).toByteArray(MsgComm.Msg.serializer())
+        )
+    }
 }
 
 @Serializable(OnlineMessageSourceFromTempImpl.Serializer::class)
 internal class OnlineMessageSourceFromTempImpl(
     override val bot: Bot,
-    private val msg: List<MsgComm.Msg>
+    msg: List<MsgComm.Msg>
 ) : OnlineMessageSource.Incoming.FromTemp(), MessageSourceInternal {
     object Serializer : MessageSourceSerializerImpl("OnlineMessageSourceFromTemp")
 
-    override val sequenceIds: IntArray get() = msg.mapToIntArray { it.msgHead.msgSeq }
-    override val internalIds: IntArray get() = msg.mapToIntArray { it.msgBody.richText.attr!!.random }
+    override val sequenceIds: IntArray = msg.mapToIntArray { it.msgHead.msgSeq }
+    override val internalIds: IntArray = msg.mapToIntArray { it.msgBody.richText.attr!!.random }
     override var isRecalledOrPlanned: AtomicBoolean = AtomicBoolean(false)
     override val ids: IntArray get() = sequenceIds//
-    override val time: Int get() = msg.first().msgHead.msgTime
+    override val time: Int = msg.first().msgHead.msgTime
     override val originalMessage: MessageChain by lazy {
-        msg.toMessageChain(
-            bot,
-            bot.id,
-            groupIdOrZero = 0,
-            onlineSource = null,
-            MessageSourceKind.TEMP
-        )
+        msg.toMessageChainNoSource(bot.id, groupIdOrZero = 0, MessageSourceKind.TEMP)
     }
-    override val sender: Member
-        get() = with(msg.first().msgHead) {
-            bot.getGroupOrFail(c2cTmpMsgHead!!.groupUin).getOrFail(fromUin)
-        }
+    override val sender: Member = with(msg.first().msgHead) {
+        bot.getGroupOrFail(c2cTmpMsgHead!!.groupUin).getOrFail(fromUin)
+    }
 
     private val jceData by lazy { msg.toJceDataPrivate(internalIds) }
     override fun toJceData(): ImMsgBody.SourceMsg = jceData
 }
 
 @Serializable(OnlineMessageSourceFromGroupImpl.Serializer::class)
-internal data class OnlineMessageSourceFromGroupImpl(
+internal class OnlineMessageSourceFromGroupImpl(
     override val bot: Bot,
-    private val msg: List<MsgComm.Msg>
+    msg: List<MsgComm.Msg>
 ) : OnlineMessageSource.Incoming.FromGroup(), MessageSourceInternal {
     object Serializer : MessageSourceSerializerImpl("OnlineMessageSourceFromGroupImpl")
 
     @Transient
     override var isRecalledOrPlanned: AtomicBoolean = AtomicBoolean(false)
-    override val sequenceIds: IntArray get() = msg.mapToIntArray { it.msgHead.msgSeq }
-    override val internalIds: IntArray get() = msg.mapToIntArray { it.msgBody.richText.attr!!.random }
+    override val sequenceIds: IntArray = msg.mapToIntArray { it.msgHead.msgSeq }
+    override val internalIds: IntArray = msg.mapToIntArray { it.msgBody.richText.attr!!.random }
     override val ids: IntArray get() = sequenceIds
-    override val time: Int get() = msg.first().msgHead.msgTime
+    override val time: Int = msg.first().msgHead.msgTime
     override val originalMessage: MessageChain by lazy {
-        msg.toMessageChain(bot, bot.id, groupIdOrZero = group.id, onlineSource = null, MessageSourceKind.GROUP)
+        msg.toMessageChainNoSource(bot.id, groupIdOrZero = group.id, MessageSourceKind.GROUP)
     }
 
     override val sender: Member by lazy {
-        (bot.getGroup(
-            msg.first().msgHead.groupInfo?.groupCode
-                ?: error("cannot find groupCode for MessageSourceFromGroupImpl. msg=${msg._miraiContentToString()}")
-        ) as GroupImpl).run {
-            get(msg.first().msgHead.fromUin)
-                ?: msg.first().msgBody.richText.elems.firstOrNull { it.anonGroupMsg != null }?.run {
-                    newAnonymous(anonGroupMsg!!.anonNick.encodeToString(), anonGroupMsg.anonId.encodeToBase64())
-                }
-                ?: error("cannot find member for MessageSourceFromGroupImpl. msg=${msg._miraiContentToString()}")
+        val groupCode = msg.first().msgHead.groupInfo?.groupCode
+            ?: error("cannot find groupCode for OnlineMessageSourceFromGroupImpl. msg=${msg._miraiContentToString()}")
+
+        val group = bot.getGroup(groupCode)?.checkIsGroupImpl()
+            ?: error("cannot find group for OnlineMessageSourceFromGroupImpl. msg=${msg._miraiContentToString()}")
+
+        val member = group[msg.first().msgHead.fromUin]
+        if (member != null) return@lazy member
+
+        val anonymousInfo = msg.first().msgBody.richText.elems.firstOrNull { it.anonGroupMsg != null }
+            ?: error("cannot find member for OnlineMessageSourceFromGroupImpl. msg=${msg._miraiContentToString()}")
+
+        anonymousInfo.run {
+            group.newAnonymous(anonGroupMsg!!.anonNick.encodeToString(), anonGroupMsg.anonId.encodeToBase64())
         }
     }
 
-    override fun toJceData(): ImMsgBody.SourceMsg {
-        return ImMsgBody.SourceMsg(
+    private val jceData by lazy {
+        ImMsgBody.SourceMsg(
             origSeqs = intArrayOf(msg.first().msgHead.msgSeq),
             senderUin = msg.first().msgHead.fromUin,
             toUin = 0,
@@ -230,5 +190,9 @@ internal data class OnlineMessageSourceFromGroupImpl(
             pbReserve = EMPTY_BYTE_ARRAY,
             srcMsg = EMPTY_BYTE_ARRAY
         )
+    }
+
+    override fun toJceData(): ImMsgBody.SourceMsg {
+        return jceData
     }
 }
